Added user session hook
